Refactor to decouple JsonSchemaDiffLibrary from registry objects.
