+++ app/src/main/java/io/apicurio/registry/rules/compatibility/JsonSchemaCompatibilityChecker.java
@@ -80,7 +80,11 @@ public class JsonSchemaCompatibilityChecker implements CompatibilityChecker {
             case NONE:
                 break;
         }
-        Set<CompatibilityDifference> diffs = incompatibleDiffs.stream().map(diff -> (CompatibilityDifference) diff).collect(Collectors.toSet());
+        Set<CompatibilityDifference> diffs =
+            incompatibleDiffs
+                .stream()
+                .map(JsonSchemaCompatibilityDifference::new)
+                .collect(Collectors.toSet());
         return CompatibilityExecutionResult.incompatible(diffs);
     }
